Resolve clippy warnings for wio_terminal

+++ boards/wio_terminal/examples/snake.rs
@@ -25,14 +25,13 @@ use wio::{button_interrupt, Button, ButtonController, ButtonEvent};
 use heapless::spsc::Queue;
 
 // pseudo-random number generation
-use oorandom;
 use oorandom::Rand32;
 
 const DISPLAY_WIDTH: u32 = 320;
 const DISPLAY_HEIGHT: u32 = 240;
 const CELL_SIZE: u32 = 10;
-const GRID_WIDTH: u32 = DISPLAY_WIDTH / CELL_SIZE as u32;
-const GRID_HEIGHT: u32 = DISPLAY_HEIGHT / CELL_SIZE as u32;
+const GRID_WIDTH: u32 = DISPLAY_WIDTH / CELL_SIZE;
+const GRID_HEIGHT: u32 = DISPLAY_HEIGHT / CELL_SIZE;
 
 static mut BUTTON_CTRLR: Option<ButtonController> = None;
 static mut Q: Queue<ButtonEvent, 8> = Queue::new();
@@ -45,6 +44,7 @@ button_interrupt!(
     }
 );
 
+#[allow(clippy::empty_loop)]
 #[entry]
 fn main() -> ! {
     // Initial initializations
